+++ crates/iceberg/src/lib.rs
@@ -99,6 +99,8 @@ pub use runtime::{JoinHandle, Runtime, RuntimeHandle};
 pub mod arrow;
 pub(crate) mod delete_file_index;
 pub mod encryption;
+#[cfg(any(test, feature = "test-utils"))]
+#[doc(hidden)]
 pub mod test_utils;
 pub mod writer;
