+++ crates/integrations/datafusion/Cargo.toml
@@ -42,6 +42,7 @@ uuid = { workspace = true }
 
 [dev-dependencies]
 expect-test = { workspace = true }
+iceberg = { workspace = true, features = ["test-utils"] }
 parquet = { workspace = true }
 tempfile = { workspace = true }
